+++ src/main/typescript/constants.ts
@@ -540,6 +540,7 @@ export const HANDLEBARS = {
       equipment: `${CONSTANTS.systemPath}/handlebars/actors/parts/equipment.hbs`,
       header: `${CONSTANTS.systemPath}/handlebars/actors/parts/header.hbs`,
       inventory: `${CONSTANTS.systemPath}/handlebars/actors/parts/inventory.hbs`,
+      inventoryEquipmentSlot: `${CONSTANTS.systemPath}/handlebars/actors/parts/inventoryEquipmentSlot.hbs`,
       magic: `${CONSTANTS.systemPath}/handlebars/actors/parts/magic.hbs`,
       stats: `${CONSTANTS.systemPath}/handlebars/actors/parts/stats.hbs`,
       weaponSlot: `${CONSTANTS.systemPath}/handlebars/actors/parts/weaponSlot.hbs`